requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
opencv-python-headless==4.8.0.76
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0
python-multipart==0.0.9
sqlalchemy==2.0.30
aiosqlite==0.20.0
insightface==0.7.3
onnxruntime==1.16.3
scipy>=1.10.0